convert heading between gopal track, gpx, columbus
git-svn-id: http://www.routeconverter.de/svn/RouteConverter/trunk@1461 964dd122-5732-0410-82e6-8cc08e73365a

+++ navigation-formats/src/java/slash/navigation/gopal/GoPalTrackFormat.java
@@ -87,8 +87,8 @@ public class GoPalTrackFormat extends SimpleLineBasedFormat<SimpleRoute> {
         Matcher matcher = LINE_PATTERN.matcher(line);
         if (!matcher.matches())
             return false;
-        Integer satellites = Conversion.parseInt(matcher.group(5));
-        return satellites != null && satellites > 0;
+        Double satellites = Conversion.parseDouble(matcher.group(5));
+        return satellites != null && satellites > 0.0;
     }
 
     private CompactCalendar parseTime(String time) {
